Fixin' n' trixin'

+++ word2vec/app/preprocessing/cleaning/data_cleaner.py
@@ -4,12 +4,11 @@ from os import path, listdir
 
 
 class DataCleaner(object):
-    def __init__(self, source_dir):
+    def __init__(self, source_dir, target_dir, dry_run=False):
         self.dir_name = path.dirname(__file__)
-        self.source_dir = path.join(self.dir_name, "../../..", source_dir)
-        self.target_dir = path.join(
-            self.dir_name, "../../../data/3_cleaned", path.basename(source_dir)
-        )
+        self.source_dir = path.join(self.dir_name, source_dir)
+        self.target_dir = path.join(self.dir_name, target_dir)
+        self.dry_run = dry_run
         stop_words_file = path.join(self.dir_name, "stop_words.txt")
         with open(stop_words_file, "r") as f:
             self.stop_words = [word for line in f for word in line.split()]
@@ -90,7 +89,7 @@ class DataCleaner(object):
         a_line = re.sub(r"[-]+[\s]+", " ", a_line)  # hyphen followed by space(s)
         # remove special characters, punctuation etc.
         a_line = re.sub(
-            r'[()‘"”…#@/&%;:`\*\'{}+_\u200B\u2013=~§\$|.!?,\[\]\\]', "", a_line
+            r'[()<>‘"”…#@/&%;:`\*\'{}+_\u200B\u2013=~§\$|.!?,\[\]\\]', "", a_line
         )
         # let's remove superfluous whitespaces, again (cleaning can have resulted in additional spaces between words)
         a_line = re.sub(" +", " ", a_line)
@@ -109,6 +108,9 @@ class DataCleaner(object):
                 for dirty_line in dirty_data:
                     cleaned_lines.append(self.clean_line(dirty_line))
 
+            if self.dry_run:
+                return cleaned_lines
+
             with open(cleaned_file, "w") as clean_data:
                 for cleaned_line in cleaned_lines:
                     # If line is shorter than 3 words we ignore it as it doesn't give much to train on
@@ -122,7 +124,8 @@ class DataCleaner(object):
 
 def main():
     clean_directory = sys.argv[1]
-    data_cleaner = DataCleaner(clean_directory)
+    target_dir = path.join("../../../data/3_cleaned", path.basename(clean_directory))
+    data_cleaner = DataCleaner(source_dir=clean_directory, target_dir=target_dir)
     data_cleaner.clean_files()
