+++ word2vec/app/preprocessing/parsing/blog_parser.py
@@ -1,12 +1,38 @@
 import yaml
 import re
 from os import path, listdir
+import logging
 
 
 class BlogParser(object):
-    def __init__(self, blog_dir):
+    logging.basicConfig(level=logging.INFO)
+
+    def __init__(self, blog_dir, dry_run=False):
         dir_name = path.dirname(__file__)
+        self.logger = logging.getLogger(__name__)
         self.blog_directory = path.join(dir_name, blog_dir)
+        self.dry_run = dry_run
+
+    def _tag_remover(self, line):
+        opening_tag_index = line.find('<')
+        closing_tag_index = line.find('>')
+        if 0 <= opening_tag_index < closing_tag_index:
+            if opening_tag_index > 0:
+                return self._tag_remover(line[:opening_tag_index - 1] + line[closing_tag_index + 1:])
+            else:
+                return self._tag_remover(line[closing_tag_index + 1:])
+        return line
+
+    def _link_remover(self, line):
+        open_bracket_index = line.find('[')
+        close_bracket_index = line.find(']')
+        if 0 <= open_bracket_index < close_bracket_index:
+            open_parenthesis_index = line[close_bracket_index:].find('(')
+            close_parenthesis_index = line[close_bracket_index:].find(')')
+            if 0 <= open_parenthesis_index < close_parenthesis_index:
+                return self._link_remover(line[:close_bracket_index + open_parenthesis_index] +
+                                          line[close_bracket_index + close_parenthesis_index + 1:])
+        return line
 
     def parse_blog_files(self):
         dir_name = path.dirname(__file__)
@@ -21,9 +47,10 @@ class BlogParser(object):
     def parse_blog_file(self, source_file, target_file):
         include_line = True
         blog_file = path.join(self.blog_directory, source_file)
+        self.logger.info(f"Parsing blog file: {blog_file}")
         with open(blog_file, "r", encoding="utf-8", errors="ignore") as blog_data:
             blog_lines = (line.rstrip() for line in blog_data if line)
-            cleaned_blog_lines = []
+            parsed_blog_lines = []
             for blog_line in blog_lines:
                 is_comment_code_block_sep = re.match("^(`|-|~){3}.*$", blog_line)
                 if is_comment_code_block_sep and include_line:
@@ -31,23 +58,21 @@ class BlogParser(object):
                 elif is_comment_code_block_sep and not include_line:
                     include_line = True
                     continue
-                if re.match("^#+.*$", blog_line):
-                    continue
                 if include_line:
                     # replace NBSP with normal space
                     blog_line = " ".join(blog_line.split())
-                    # remove anything within parenthesis, usually url:s
-                    blog_line = re.sub(r"\(.*\)", "", blog_line)
-                    # remove some markup code
-                    blog_line = re.sub(r"\{.*\}", "", blog_line)
-                    # remove quoted words
-                    blog_line = re.sub(r"`[^\s]*`", "", blog_line)
-                    cleaned_blog_lines.append(blog_line)
-            blog_as_string = " ".join(cleaned_blog_lines)
+                    # remove links:
+                    blog_line = self._link_remover(blog_line)
+                    # remove html tags
+                    blog_line = self._tag_remover(blog_line)
+                    # remove quotes and stars, parenthesis and brackets
+                    # blog_line = re.sub(r"`\"\'\*\[\]\(\)", "", blog_line)
+                    parsed_blog_lines.append(blog_line)
+            if self.dry_run:
+                return parsed_blog_lines
             with open(target_file, "w") as prepared_data:
-                # write each sentence on a single line
-                for blog_sentence in blog_as_string.split("."):
-                    prepared_data.write(blog_sentence.strip())
+                for parsed_blog_line in parsed_blog_lines:
+                    prepared_data.write(parsed_blog_line.strip())
                     prepared_data.write("\n")
